aider: Added a `BottomNavigationBar` to the home screen with two items: Contacts and Job Card, and implemented navigation to the respective screens when the items are selected.
# Aider chat conversation:

USER: All i see it the contacts

ASSISTANT: Currently, the home screen of your application is displaying a list of contacts fetched from the `res.partner` model. If you want to navigate to the `JobCardCreate` screen from the home screen, you need to add a button or a menu item that will trigger the navigation.

Let's add a floating action button to the home screen that will navigate to the `JobCardCreate` screen when pressed. We will need to modify the `lib\src\home\view\home.dart` file.

Here is the *edit block* for the changes:

```dart
lib\src\home\view\home.dart
<<<<<<< HEAD
    return MainContainer(

^C KeyboardInterrupt

USER: I want to have a new windows one whoch says contacts whoch brings to Contacts anotehr whoch says  job card when click it bring you there

ASSISTANT: To achieve this, we can use a `BottomNavigationBar` widget in your `Home` widget. This will allow us to switch between different screens from the home screen. We will create two items in the `BottomNavigationBar`: one for Contacts and another for Job Card. When the Contacts item is selected, it will display the list of contacts. When the Job Card item is selected, it will navigate to the `JobCardCreate` screen.

Here is the *edit block* for the changes:

```dart
lib\src\home\view\home.dart
<<<<<<< HEAD
class _HomeState extends State<Home> {
  final _homeController = Get.put(HomeController());

  @override
  void initState() {
    super.initState();
    _homeController.getPartners();
  }

  @override
  Widget build(BuildContext context) {
    return MainContainer(
      isAppBar: true,
      appBarTitle: "Home",
      padding: 20.0,
      actions: [
        IconButton(
          onPressed: () {
            showLogoutDialog();
          },
          icon: Icon(Icons.exit_to_app),
        )
      ],
      child: Obx(
        () {
          return ListView.builder(
            shrinkWrap: true,
            physics: ScrollPhysics(),
            itemCount: _homeController.listOfPartners.length,
            itemBuilder: (context, index) {
              return ListTile(
                leading: ClipRRect(
                  borderRadius: BorderRadius.circular(8.0),
                  child: _homeController
                          .listOfPartners[index].image512!.isNotEmpty
                      ? Image.memory(
                          base64.decode(
                              _homeController.listOfPartners[index].image512!),
                          height: 40,
                          width: 40,
                        )
                      : Icon(
                          Icons.person,
                          color: AppColors.grey,
                          size: 40,
                        ),
                ),
                title: Text(_homeController.listOfPartners[index].name!),
                subtitle: Text(_homeController.listOfPartners[index].email!),
              );
            },
          );
        },
      ),
    );
  }
}
=======
class _HomeState extends State<Home> {
  final _homeController = Get.put(HomeController());
  int _currentIndex = 0;

  @override
  void initState() {
    super.initState();
    _homeController.getPartners();
  }

  @override
  Widget build(BuildContext context) {
    return Scaffold(
      appBar: AppBar(
        title: Text('Home'),
        actions: [
          IconButton(
            onPressed: () {
              showLogoutDialog();
            },
            icon: Icon(Icons.exit_to_app),
          ),
        ],
      ),
      body: _currentIndex == 0 ? _buildContacts() : JobCardCreate(),
      bottomNavigationBar: BottomNavigationBar(
        currentIndex: _currentIndex,
        items: [
          BottomNavigationBarItem(
            icon: Icon(Icons.contacts),
            label: 'Contacts',
          ),
          BottomNavigationBarItem(
            icon: Icon(Icons.work),
            label: 'Job Card',
          ),
        ],
        onTap: (index) {
          setState(() {
            _currentIndex = index;
          });
        },
      ),
    );
  }

  Widget _buildContacts() {
    return Obx(
      () {
        return ListView.builder(
          shrinkWrap: true,
          physics: ScrollPhysics(),
          itemCount: _homeController.listOfPartners.length,
          itemBuilder: (context, index) {
            return ListTile(
              leading: ClipRRect(
                borderRadius: BorderRadius.circular(8.0),
                child: _homeController
                        .listOfPartners[index].image512!.isNotEmpty
                    ? Image.memory(
                        base64.decode(
                            _homeController.listOfPartners[index].image512!),
                        height: 40,
                        width: 40,
                      )
                    : Icon(
                        Icons.person,
                        color: AppColors.grey,
                        size: 40,
                      ),
              ),
              title: Text(_homeController.listOfPartners[index].name!),
              subtitle: Text(_homeController.listOfPartners[index].email!),
            );
          },
        );
      },
    );
  }
}
>>>>>>> updated
```

In this change, we replaced the `MainContainer` widget with a `Scaffold` widget that contains a `BottomNavigationBar`. The `BottomNavigationBar` has two items: Contacts and Job Card. We also added a `_currentIndex` variable to keep track of the currently selected item in the `BottomNavigationBar`. The body of the `Scaffold` changes based on the value of `_currentIndex`. If `_currentIndex` is 0, it displays the contacts list. If `_currentIndex` is 1, it navigates to the `JobCardCreate` screen.

Please note that you need to import the `JobCardCreate` widget at the top of the `home.dart` file. If the import statement is not already there, add this line at the top of the file:

```dart
import 'package:your_package_name/src/job_card/view/job_card_create.dart';
```

Replace `your_package_name` with the actual name of your package.