aider: Fixed 'static' modifier error in createJobCardWithImage method.
# Aider chat conversation:

USER: [{
"resource": "/D:/Repos/Odoo-JsonRpc-with-Flutter/lib/common/api_factory/api.dart",
"owner": "_generated_diagnostic_collection_name_#0",
"code": "extraneous_modifier",
"severity": 8,
"message": "Can't have modifier 'static' here.\nTry removing 'static'.",
"source": "dart",
"startLineNumber": 449,
"startColumn": 3,
"endLineNumber": 449,
"endColumn": 9
}]

ASSISTANT: The error message indicates that the `static` modifier is not allowed in the location where it's currently used. In Dart, `static` is used to denote that a member variable, or method, belongs to a class itself, rather than to instances of the class.

The `static` keyword is not allowed in the body of a method or function. It seems like the `createJobCardWithImage` function is defined inside another function or method, which is not allowed.

Here is the corrected code:

```dart
lib\common\api_factory\api.dart
<<<<<<< HEAD
  static createJobCardWithImage(String base64Image, String customerId, String regNo) {
    // TODO: Implement the API call to send the image, customer ID, and registration number to the Odoo API
  }
=======
  static Future<void> createJobCardWithImage(String base64Image, String customerId, String regNo) async {
    // TODO: Implement the API call to send the image, customer ID, and registration number to the Odoo API
  }
>>>>>>> updated
```

Please replace the `TODO` comment with the actual implementation of the API call. The implementation will depend on the specifics of your API, such as the endpoint, the HTTP method (GET, POST, etc.), and the format of the request body.