aider: (no commit message provided)
# Aider chat conversation:

USER: what is th epurpose of the avatar url

ASSISTANT: The `avatarUrl` is typically used to store the URL of a user's profile picture or avatar. In the context of your application, it seems to be used to display an image in the `JobCardCreate` form. However, it's not clear how this `avatarUrl` is used in the rest of your application. If it's not being used, you might consider removing it to simplify your code.

USER: i dont need to see this in the flutter ap. can I not take an image and when we click submit it send to odoo. Now when i taje an image it does not pass anything to the url it is nt doeing aything

ASSISTANT: Yes, you can modify your application to send the image directly to Odoo when the submit button is clicked, without displaying the avatar URL in the Flutter app.

To do this, you need to modify the `onPressed` function of the `ElevatedButton` in the `JobCardCreate` class to send the image directly to the Odoo API when the submit button is clicked.

Here are the changes you need to make:

1. Remove the `TextFormField` and `Container` widgets related to the `avatarUrlController` in the `JobCardCreate` class.

2. Modify the `onPressed` function of the `ElevatedButton` in the `JobCardCreate` class to call the `createJobCardApi` function instead of the `createJobCardWithImage` function.

Here are the *edit blocks* for these changes:

```dart
lib\src\job_card\view\job_card_create.dart
<<<<<<< HEAD
              TextFormField(
                controller: _avatarUrlController,
                decoration: InputDecoration(labelText: 'Avatar URL'),
                validator: (value) {
                  if (value == null || value.isEmpty) {
                    return 'Please enter an avatar URL';
                  }
                  return null;
                },
              ),
              Container(
                child: _avatarUrlController.text.isEmpty
                    ? Text('No image')
                    : Image.network(_avatarUrlController.text),
              ),
              ElevatedButton(
                onPressed: () async {
                  if (_formKey.currentState!.validate()) {
                    String base64Image = base64Encode(_image!.readAsBytesSync());
                    await Api.createJobCardWithImage(base64Image, _customerIdController.text, _regNoController.text);
                  }
                },
                child: Text('Submit'),
              ),
=======
              ElevatedButton(
                onPressed: () async {
                  if (_formKey.currentState!.validate()) {
                    String base64Image = base64Encode(_image!.readAsBytesSync());
                    createJobCardApi(
                      base64Image: base64Image,
                      customerId: _customerIdController.text,
                      regNo: _regNoController.text,
                      onSuccess: () {
                        print('Job card created successfully');
                      },
                      onError: () {
                        print('Failed to create job card');
                      },
                    );
                  }
                },
                child: Text('Submit'),
              ),
>>>>>>> updated
```

Please note that you need to import the `createJobCardApi` function at the top of the `job_card_create.dart` file:

```dart
lib\src\job_card\view\job_card_create.dart
<<<<<<< HEAD
import 'package:odoo_common_code_latest/common/api_factory/api.dart';
=======
import 'package:odoo_common_code_latest/common/api_factory/api.dart';
import 'package:odoo_common_code_latest/common/api_factory/modules/job_card_api_module.dart';
>>>>>>> updated
```

After these changes, the image will be sent directly to the Odoo API when the submit button is clicked, without displaying the avatar URL in the Flutter app.