aider: (no commit message provided)
# Aider chat conversation:

USER: i get this in the odoo log File "/usr/lib/python3/dist-packages/odoo/api.py", line 410, in _model_create_multi
    return create(self, arg)
  File "/usr/lib/python3/dist-packages/odoo/models.py", line 3892, in create
    raise ValueError("Invalid field %r on model %r" % (key, self._name))
ValueError: Invalid field 'base64_image' on model 'job.card'
2023-08-12 01:46:14,377 1 INFO car werkzeug: 192.168.20.190 - - [12/Aug/2023 01:46:14] "POST /web/dataset/call_kw/job.card/create HTTP/1.1" 200 - 6 0.004 0.171

ASSISTANT: The error message indicates that the field `base64_image` is not a valid field on the `job.card` model in your Odoo instance.

You need to ensure that the `job.card` model in your Odoo instance has a field named `base64_image`. If it doesn't, you need to add this field to the model.

If the field exists but has a different name, you need to replace `base64_image` with the correct field name in the `createJobCardApi` function in the `job_card_api_module.dart` file.

If you confirm the correct field name, I can provide the *edit block* for the change.

USER: the feild name is avatar

ASSISTANT: Great, we will replace `base64_image` with `avatar` in the `createJobCardApi` function in the `job_card_api_module.dart` file.

Here is the *edit block* for the change:

```dart
lib\common\api_factory\modules\job_card_api_module.dart
<<<<<<< HEAD
    values: {
      "base64_image": base64Image,
      "customer_id": customerId,
      "reg_no": regNo,
    },
=======
    values: {
      "avatar": base64Image,
      "customer_id": customerId,
      "reg_no": regNo,
    },
>>>>>>> updated
```

This change replaces the key `base64_image` with `avatar` in the `values` map.