aider: (no commit message provided)
# Aider chat conversation:

USER: the make feild should be vehicle_brand_id

ASSISTANT: To change the 'make' field to 'vehicle_brand_id', we need to modify the `lib\src\job_card\view\job_card_create.dart` file.

Here is the *edit block* for the change:

```dart
lib\src\job_card\view\job_card_create.dart
<<<<<<< HEAD
                  value: _selectedVehicleMake,
                  items: _vehicleMakes.map((make) {
                    return DropdownMenuItem(
                      value: make,
                      child: Text(make),
=======
                  value: _selectedVehicleBrandId,
                  items: _vehicleBrandIds.map((brandId) {
                    return DropdownMenuItem(
                      value: brandId,
                      child: Text(brandId.toString()),
>>>>>>> updated
                    );
                  }).toList(),
                  onChanged: (value) {
                    setState(() {
                      _selectedVehicleBrandId = value as int?;
                    });
                  },
                  decoration: InputDecoration(labelText: 'Vehicle Brand ID'),
                ),
```

Please note that this change assumes that `_selectedVehicleBrandId` and `_vehicleBrandIds` are defined in your code and `_vehicleBrandIds` is a list of integers. If this is not the case, you will need to define these variables and populate `_vehicleBrandIds` with the appropriate data.