aider: (no commit message provided)
# Aider chat conversation:

USER:
  File "/usr/lib/python3/dist-packages/odoo/api.py", line 410, in _model_create_multi
    return create(self, arg)
  File "/usr/lib/python3/dist-packages/odoo/addons/base/models/ir_fields.py", line 665, in create
    recs = super().create(vals_list)
  File "<decorator-gen-13>", line 2, in create
  File "/usr/lib/python3/dist-packages/odoo/api.py", line 410, in _model_create_multi
    return create(self, arg)
  File "/usr/lib/python3/dist-packages/odoo/models.py", line 3892, in create
    raise ValueError("Invalid field %r on model %r" % (key, self._name))
ValueError: Invalid field 'vehicle_make' on model 'job.card'
2023-08-12 18:56:24,890 1 INFO car werkzeug: 192.168.20.190 - - [12/Aug/2023 18:56:24] "POST /web/dataset/call_kw/job.card/create HTTP/1.1" 200 - 6 0.005 0.066

ASSISTANT: The error message indicates that the field 'vehicle_make' is not a valid field on the 'job.card' model in your Odoo database. This could be due to a typo in the field name or the field might not exist in the 'job.card' model.

To resolve this issue, you need to ensure that the field 'vehicle_make' exists in the 'job.card' model in your Odoo database. If the field exists but has a different name, you should replace 'vehicle_make' with the correct field name in your code.

The code that needs to be modified is in the `lib\common\api_factory\modules
^C KeyboardInterrupt

USER: vehicle_make' is wrong frils this doesn not exsist it should be vehicle_brand_id this was a typo by you can you fix

ASSISTANT: Sure, I can help with that. We need to replace 'vehicle_make' with 'vehicle_brand_id' in the `createJobCardApi` function in the `lib\common\api_factory\modules\job_card_api_module.dart` file.

Here is the *edit block* for the change:

```dart
lib\common\api_factory\modules\job_card_api_module.dart
<<<<<<< HEAD
createJobCardApi({
  required String base64Image,
  required String customerId,
  required String regNo,
  required String vehicleMake,
  required String vehicleModel,
  required int vehicleBrandId,
  required int vehicleModelId,
  required Function onSuccess,
  required Function onError,
}) {
  Api.create(
    model: "job.card",
    values: {
      "avatar": base64Image,
      "customer_id": customerId,
      "reg_no": regNo,
      "vehicle_make": vehicleMake,
      "vehicle_model": vehicleModel,
      "vehicle_brand_id": vehicleBrandId,
      "vehicle_model_id": vehicleModelId,
    },
    onResponse: (response) {
      print('Job card created with id: ${response}');
      onSuccess();
    },
    onError: (error, data) {
      handleApiError(error);
      onError();
    },
  );
}
=======
createJobCardApi({
  required String base64Image,
  required String customerId,
  required String regNo,
  required String vehicleMake,
  required String vehicleModel,
  required int vehicleBrandId,
  required int vehicleModelId,
  required Function onSuccess,
  required Function onError,
}) {
  Api.create(
    model: "job.card",
    values: {
      "avatar": base64Image,
      "customer_id": customerId,
      "reg_no": regNo,
      "vehicle_brand_id": vehicleBrandId,
      "vehicle_model": vehicleModel,
      "vehicle_brand_id": vehicleBrandId,
      "vehicle_model_id": vehicleModelId,
    },
    onResponse: (response) {
      print('Job card created with id: ${response}');
      onSuccess();
    },
    onError: (error, data) {
      handleApiError(error);
      onError();
    },
  );
}
>>>>>>> updated
```

This *edit block* replaces the 'vehicle_make' field with 'vehicle_brand_id' in the `createJobCardApi` function.